move math_util to lib
