+++ slm_lab/lib/math_util.py
